requirements.txt
Flask>=2.0
Flask-SocketIO>=5.1
pymongo>=3.11
python-dotenv>=0.19.2
eventlet>=0.33.0   # (if using eventlet with Flask-SocketIO for async support)
